date: 1999/07/05
Add ELSEIF to *FORTRAN-KEYWORDS*.

In check_new_vbles, we forgot to skip over strings, FREF-STRING, and
|:| (the substring indexing character)

In lookup-vble-type, we didn't return the correct type if the array
was declared only using a DIMENSION statement.  In this case we need
to use Fortran implicit rules to get the type.

In GET-FUN-ARG-TYPE, VBLE-DECLARED-TWICE-P doesn't always return a
list of two elements.

In FIXUP-EXPRESSION, we forgot to include the rest of the expression
during fixup, in certain cases.