date: 1999/07/06
Fortran PARAMETER variables are no longer keyword arguments to the
function.  Instead, a closure is created with the PARAMETER variables
appropriately set.  This means you can't dynamically change PARAMETER
values; you need to edit the source and recompile.

This change was needed because SAVE'd variables can create closures,
and if the SAVE'd variables use the PARAMETER variables, as
dimensions, say, then, the PARAMETER value must be available to the
SAVE'd variables.