First pass at creating common blocks as arrays.  Intent is to allow odepack to be converted via f2cl.
So a common block structure is created that creates as large an array
as possible for consecutive elements of the same type in the common
block.  A new array is created for each such section.  Then the
elements of the common block are accessed either as either an
individual element of the array or as a displaced array.

This might have speed impacts, so the default is not to do this.  Use
the keyword :common-as-array to control this feature.  Default is off,
preserving old behavior.